Support for automatically handle help options
Add `handle-help?' keyword parameter (default: `#t') to
`define-command'.  This parameter causes help options (value of the
`help-options' parameter) to be implicitly intercepted and handled,
causing the help messge provided at command definition to be printed
to stdout and the program to be terminated, exiting 0.

`command-usage' has been renamed to `show-command-help' for
consistency.